Add config_show4GForLte false to US carrier config.xml
By default we now show 4G instead of LTE by changing
config_show4GForLte to true in
  frameworks/base/packages/SystemUI/res/values/config.xml

So for US carriers we are adding a config.xml which has
config_show4GForLte to false. The list of overlays comes from
looking for mcc=310 & mcc=311 in the apns-conf.xml for vendor deb.

Bug: 9265217
Change-Id: I962910de7826ce9825de2d2dfb9e50b84d0e7c75