Enable hspa being distinguisable for Deb.
The SystemUI resource config_hspa_data_distinguishable was not defined
for Deb. I copied the file from Flo to Deb so its now enabled and the
bug fixed.

This file also includes config_search_panel_view_vibration_duration.
I thought it best for Deb and Flo to be the same so I've left that in
and the files are identical for Deb and Flo.

Bug: 9747103
Change-Id: I5f393cfadd06450812424e55c5583811a52527ce